requirements.txt
fastapi==0.110.2
langchain==0.1.16
langchain_community==0.0.34
langchain_core==0.1.45
langchain_openai==0.1.3
pydantic==2.7.0
PyJWT==2.4.0
PyJWT==2.8.0
PyMySQL==1.0.2
Requests==2.31.0
torch==2.2.2
transformers==4.40.0
